getValOrDefault → getFirstVal
The original getValOrDefault was a binary utility function that made it easy to say "return the first value I gave you, unless it's undefined, null, or a null string; if that's the case, return the second default value." But with defaults set by environment variables that might or might not be set, we have a use case for a three-way choice: the variable, the environment variable default, or a hard default if no environment variable is chosen. So this lets you use it to chain together as many things as you like, and it will return the first thing it gets that isn't undefined, null, or a null string. If you pass it nothing but those, it returns undefined. This seemed like a sufficiently significant change to warrant a name change, and that meant updating all the places that it's used.