checkAny, checkAll: Add short-circuiting
This lets the method return as soon as it has an answer — for example, the first time it finds a false in checkAll, or the first time it finds a true in checkAny. This lets us use short-circuiting like we normally would, e.g., asking if a variable is undefined or if it's an array with at least three elements. If it's undefined, we bail out and get our answer, and never ask about the length of our undefined variable, which would throw an error.