createUser middleware: Remove User.save
If we want more atomic middlewares, we might be doing a few things with this user, and we wouldn't want to be running back to the database for each and every one. So, in the spirit of more atomic middlewares, let's take the saving out and make a separate middleware that's just for saving the subject.