Parser: parseKeyValPair: Allow for object or uniform object
Well that's clear as mud. What I mean is you can either get an object that reflects the key/value pair (e.g., { key=value }), or one with a key property that's the key and a value property that's the value. There's a time and a place for each one.