Add expecter: isUser
Our integration tests end up testing for very similar things under a bunch of slightly different conditions, meaning that we're repeating a lot of the same expectations over and over again. Turns out you can put assertions in a separate function and they run just fine! So we're going to start grouping these repeated assertions into "expecter" methods to make our test files a little shorter and easier to manage. This is the first one, which tests if the thing it's given is a user record that matches the expected value provided.